corrected auth(add policies)

+++ resources/views/labels/index.blade.php
@@ -28,21 +28,20 @@
                 <td>{{ $label->name }}</td>
                 <td>{{ $label->description }}</td>
                 <td>{{ $label->created_at->format('d.m.Y') }}</td>
-                @auth
-                    <td>
-                        <a
-                            class="text-danger text-decoration-none"
+                <td>
+                    @can('delete', $label)
+                        <a  class="text-danger text-decoration-none"
                             href="{{ route('labels.destroy', $label) }}"
                             data-confirm="{{ __('messages.are_you_sure') }}"
                             data-method="delete"
-                            rel="nofollow"
-                        >
+                            rel="nofollow">
                             {{ __('messages.to_delete') }}                        </a>
-
+                    @endcan
+                    @can('update', $label)
                         <a class="text-decoration-none" href="{{ route('labels.edit', $label) }}">
                             {{ __('messages.to_change') }}                        </a>
-                    </td>
-                @endauth
+                    @endcan
+                </td>
             </tr>
 
         @endforeach
